scripts/midocheck/requirements.txt
Click>=6.0
midonetclient
openstacksdk
prettytable
